requirements.txt
Django==1.9.1
django-extensions==1.6.7
djangorestframework==3.2.3
psycopg2==2.6.1
six==1.11.0
model-mommy==1.2.6
